pyproject.toml
[tool.poetry]
name = "petstoreanduseinsider"
version = "0.1.0"
description = ""
authors = ["Bohdan Vinichenkpo <seinzumliebe@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.29.0"
pytest = "^7.3.1"
selenium = "^4.9.0"
webdriver-manager = "^3.8.6"
pytest-xdist = "^3.2.1"
pytest-html = "^3.2.0"
update = "^0.0.1"
allure-pytest = "^2.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
